rename file and package
